+++ o8g/Scripts/meta.py
@@ -373,6 +373,18 @@ def clearAttachLinks(card):
    setGlobalVariable('Host Cards',str(hostCards))
    debugNotify("<<< clearAttachLinks()", 3) #Debug   
 
+def sendToTrash(card, pile = None): # A function which takes care of sending a card to the right trash pile and running the appropriate scripts. Doesn't handle costs.
+   debugNotify(">>> sendToTrash()") #Debug   
+   if not pile: pile = card.owner.piles['Heap/Archives(Face-up)'] # I can't pass it as a function variable. OCTGN doesn't like it.
+   executePlayScripts(card,'TRASH') # We don't want to run automations on simply revealed cards.
+   autoscriptOtherPlayers('CardTrashed',card)
+   clearAttachLinks(card)
+   if chkModulator(card, 'preventTrash', 'onTrash'): # IF the card has the preventTrash modulator, it's not supposed to be trashed.
+      if chkModulator(card, 'ifAccessed', 'onTrash') and ds != 'runner': card.moveTo(pile) # Unless it only has that modulator active during runner access. Then when the corp trashes it, it should trash normally.
+   else: card.moveTo(pile)
+   debugNotify("<<< sendToTrash()", 3) #Debug   
+   
+   
 def resetAll(): # Clears all the global variables in order to start a new game.
    global Stored_Name, Stored_Type, Stored_Cost, Stored_Keywords, Stored_AutoActions, Stored_AutoScripts
    global installedCount, debugVerbosity, newturn,endofturn, currClicks, turn, autoRezFlags